requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blis==0.7.11
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
confection==0.1.3
cymem==2.0.8
easyocr==1.7.1
EasyProcess==1.1
elastic-transport==8.4.0
elasticsearch==8.10.0
entrypoint2==1.1
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.9.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.2
idna==3.4
imageio==2.31.3
Jinja2==3.1.2
langcodes==3.3.0
lazy-loader==0.3
lingua-language-detector==1.3.2
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
networkx==3.1
ninja==1.11.1
numpy==1.26.0
opencv-python-headless==4.8.0.76
packaging==23.1
pathy==0.10.2
patool==1.15
Pillow==10.0.1
pip==23.2.1
preshed==3.0.9
pyclipper==1.3.0.post5
pydantic-core==2.6.3
pydantic==2.3.0
pytesseract==0.3.10
python-bidi==0.4.2
python-dotenv==1.0.0
python-magic==0.4.27
pyunpack==0.3
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2022.10.31
requests==2.31.0
safetensors==0.3.3
scikit-image==0.21.0
scipy==1.11.2
setuptools==65.5.1
shapely==2.0.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy-alignments==0.9.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.2.5
spacy==3.6.1
srsly==2.4.8
sympy==1.12
thinc==8.1.12
tifffile==2023.9.18
tika==2.6.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.30.2
typer==0.9.0
typing-extensions==4.8.0
ufal.morphodita==1.11.1.1
urllib3==1.26.16
wasabi==1.1.2
wheel==0.38.4
yarl==1.9.2
